refactor: Extract collection query options to separate module
Move collection query options logic from useCollection hook to a dedicated options module for improved modularity and separation of concerns
